requirement.txt
appium-python-client==3.1.0
selenium==4.14.0
pytest==8.2.1
allure-pytest==2.13.2